# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Work Agent",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/documents": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Upload",
        "operationId": "upload_api_v1_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_v1_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Create Note",
        "operationId": "create_note_api_v1_knowledge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/{item_id}": {
      "put": {
        "tags": [
          "knowledge"
        ],
        "summary": "Update Note",
        "operationId": "update_note_api_v1_knowledge__item_id__put",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/{item_id}/versions": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Versions",
        "operationId": "versions_api_v1_knowledge__item_id__versions_get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeVersionView"
                  },
                  "title": "Response Versions Api V1 Knowledge  Item Id  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/{item_id}/restore": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Restore Note",
        "operationId": "restore_note_api_v1_knowledge__item_id__restore_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreVersion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/github/connections": {
      "get": {
        "tags": [
          "GitHub"
        ],
        "summary": "Connections",
        "operationId": "connections_api_v1_github_connections_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConnectionView"
                  },
                  "type": "array",
                  "title": "Response Connections Api V1 Github Connections Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "GitHub"
        ],
        "summary": "Create Connection",
        "operationId": "create_connection_api_v1_github_connections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/github/connections/{identity}": {
      "patch": {
        "tags": [
          "GitHub"
        ],
        "summary": "Update Connection",
        "operationId": "update_connection_api_v1_github_connections__identity__patch",
        "parameters": [
          {
            "name": "identity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identity"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/github/connections/{identity}/test": {
      "post": {
        "tags": [
          "GitHub"
        ],
        "summary": "Test Connection",
        "operationId": "test_connection_api_v1_github_connections__identity__test_post",
        "parameters": [
          {
            "name": "identity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionTest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/github/repositories/{identity}": {
      "patch": {
        "tags": [
          "GitHub"
        ],
        "summary": "Update Repository",
        "operationId": "update_repository_api_v1_github_repositories__identity__patch",
        "parameters": [
          {
            "name": "identity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identity"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepositoryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/github/issues": {
      "get": {
        "tags": [
          "GitHub"
        ],
        "summary": "Issues",
        "operationId": "issues_api_v1_github_issues_get",
        "parameters": [
          {
            "name": "owner",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner"
            }
          },
          {
            "name": "repository_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Repository Id"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(open|closed)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "label",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Label"
            }
          },
          {
            "name": "assignee",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assignee"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/github/issues/{identity}": {
      "get": {
        "tags": [
          "GitHub"
        ],
        "summary": "Issue Detail",
        "operationId": "issue_detail_api_v1_github_issues__identity__get",
        "parameters": [
          {
            "name": "identity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "GitHub"
        ],
        "summary": "Update Issue",
        "operationId": "update_issue_api_v1_github_issues__identity__patch",
        "parameters": [
          {
            "name": "identity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identity"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalStatusInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/github/issues/{identity}/conversation": {
      "post": {
        "tags": [
          "GitHub"
        ],
        "summary": "Link Chat",
        "operationId": "link_chat_api_v1_github_issues__identity__conversation_post",
        "parameters": [
          {
            "name": "identity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/github/sync": {
      "post": {
        "tags": [
          "GitHub"
        ],
        "summary": "Sync",
        "operationId": "sync_api_v1_github_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_api_v1_documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_api_v1_documents_post"
      },
...
      "ConnectionInput": {
        "properties": {
          "owner": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9-]{0,38}$",
            "title": "Owner"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "token": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 10,
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "owner",
          "name",
          "token"
        ],
        "title": "ConnectionInput"
      },
      "ConnectionTest": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "repositories": {
            "type": "integer",
            "title": "Repositories"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "repositories",
          "message"
        ],
        "title": "ConnectionTest"
      },
      "ConnectionUpdate": {
        "properties": {
          "token": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "type": "object",
        "title": "ConnectionUpdate"
      },
      "ConnectionView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "demo": {
            "type": "boolean",
            "title": "Demo"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "last_discovery": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Discovery"
          },
          "retry_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner",
          "name",
          "enabled",
          "demo",
          "status",
          "error",
          "last_discovery",
          "retry_at"
        ],
        "title": "ConnectionView"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "external_id": {
            "type": "string",
            "title": "External Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "sender": {
            "type": "string",
            "title": "Sender"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "thread_key": {
            "type": "string",
            "title": "Thread Key"
          },
          "web_url": {
            "type": "string",
            "title": "Web Url"
          },
          "meta": {
            "$ref": "#/components/schemas/ItemMetadata"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "processing": {
            "type": "string",
            "title": "Processing"
          },
          "processing_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Error"
          },
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "source_kind": {
            "type": "string",
            "title": "Source Kind"
          },
          "ai_enabled": {
            "type": "boolean",
            "title": "Ai Enabled"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_id",
          "external_id",
          "kind",
          "title",
          "body",
          "sender",
          "occurred_at",
          "thread_key",
          "web_url",
          "meta",
          "content_hash",
          "available",
          "status",
          "summary",
          "version",
          "processing",
          "source_name",
          "source_kind",
          "ai_enabled"
        ],
        "title": "ItemView"
      },
...
      "KnowledgeVersionView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "item_id",
          "version",
          "title",
          "content",
          "citations",
          "created_at"
        ],
        "title": "KnowledgeVersionView"
      },
...
      "NoteInput": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Title"
          },
          "content": {
            "type": "string",
            "maxLength": 100000,
            "title": "Content"
          },
          "expected_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "title": "NoteInput"
      },
...
      "RepositoryUpdate": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "ai_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Enabled"
          }
        },
        "type": "object",
        "title": "RepositoryUpdate"
      },
      "RepositoryView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "connection_id": {
            "type": "string",
            "title": "Connection Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "web_url": {
            "type": "string",
            "title": "Web Url"
          },
          "default_branch": {
            "type": "string",
            "title": "Default Branch"
          },
          "archived": {
            "type": "boolean",
            "title": "Archived"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "ai_enabled": {
            "type": "boolean",
            "title": "Ai Enabled"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "demo": {
            "type": "boolean",
            "title": "Demo"
          },
          "last_sync": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync"
          }
        },
        "type": "object",
        "required": [
          "id",
          "connection_id",
          "source_id",
          "owner",
          "name",
          "full_name",
          "web_url",
          "default_branch",
          "archived",
          "enabled",
          "ai_enabled",
          "status",
          "error",
          "demo",
          "last_sync"
        ],
        "title": "RepositoryView"
      },
      "RestoreVersion": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "expected_version": {
            "type": "integer",
            "title": "Expected Version"
          }
        },
        "type": "object",
        "required": [
          "version",
          "expected_version"
        ],
        "title": "RestoreVersion"
      },
...
    }
  }
}